Changed type parameters on KalmanState trait into associated types
Modified the type parameters on KalmanState trait (the float type (f32, f64) and state length) to be associated types instead. That allows me to, when creating an ExtendedKalmanFilter and giving the state, to not explicitly have to give the float type and state length as type parameters on the ExtendedKalmanFilter.