Added Jacobian, Covariance type aliases to make the code less noisy
I don't think I can make it less noisy, since it looks like I'll have to use "as" at least once to disambiguate traits. Still want to see if it's possible to write up a default implementation of F, W and H, V. Seems it may be tricky or not possible.